Remove use of PyCXX in core C++ extensions

+++ lib/matplotlib/axes/_base.py
@@ -2063,8 +2063,8 @@ class _AxesBase(martist.Artist):
                    for z, im in zorder_images]
 
             l, b, r, t = self.bbox.extents
-            width = mag * ((round(r) + 0.5) - (round(l) - 0.5))
-            height = mag * ((round(t) + 0.5) - (round(b) - 0.5))
+            width = int(mag * ((round(r) + 0.5) - (round(l) - 0.5)))
+            height = int(mag * ((round(t) + 0.5) - (round(b) - 0.5)))
             im = mimage.from_images(height,
                                     width,
                                     ims)
